+++ packages/nimble-components/src/rich-text/mention-listbox/template.ts
@@ -1,8 +1,10 @@
-import { html, ref, slotted } from '@microsoft/fast-element';
+import { html, ref, slotted, when } from '@microsoft/fast-element';
 import { Listbox } from '@microsoft/fast-foundation';
 import type { RichTextMentionListbox } from '.';
 import { anchoredRegionTag } from '../../anchored-region';
+import { filterNoResultsLabel } from '../../label-provider/core/label-tokens';
 
+/* eslint-disable @typescript-eslint/indent */
 // prettier-ignore
 export const template = html<RichTextMentionListbox>`
     <template>
@@ -18,7 +20,11 @@ export const template = html<RichTextMentionListbox>`
             horizontal-scaling="anchor"
             ?hidden="${x => !x.open}">
             <div
-                class="listbox"
+                class="
+                    listbox
+                    scrollable-region
+                    ${x => (x.filteredOptions.length === 0 ? 'empty' : '')}
+                "
                 part="listbox"
                 role="listbox"
                 @click="${(x, c) => x.clickHandler(c.event as MouseEvent)}"
@@ -27,9 +33,18 @@ export const template = html<RichTextMentionListbox>`
                 ${ref('listbox')}
             >
                 <slot name="option"
-                    ${slotted({ filter: (n: Node) => n instanceof HTMLElement && Listbox.slottedOptionFilter(n), flatten: true, property: 'slottedOptions' })}
+                    ${slotted({
+                        filter: (n: Node) => n instanceof HTMLElement && Listbox.slottedOptionFilter(n),
+                        flatten: true,
+                        property: 'slottedOptions'
+                    })}
                 >
                 </slot>
+                ${when(x => x.filteredOptions.length === 0, html<RichTextMentionListbox>`
+                    <span class="no-results-label">
+                        ${x => filterNoResultsLabel.getValueFor(x)}
+                    </span>
+                `)}
             </div>
         </${anchoredRegionTag}>
     </template>