+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,65 @@
 # Change Log - @ni/spright-components
 
-This log was last generated on Mon, 22 Jul 2024 15:45:21 GMT and should not be manually modified.
+This log was last generated on Tue, 30 Jul 2024 20:21:00 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.0.15
+
+Tue, 30 Jul 2024 20:21:00 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.6
+
+## 2.0.14
+
+Tue, 30 Jul 2024 18:17:01 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.5
+
+## 2.0.13
+
+Tue, 30 Jul 2024 16:53:07 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.4
+
+## 2.0.12
+
+Tue, 30 Jul 2024 14:34:23 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.3
+- Bump @ni/nimble-tokens to v7.1.0
+
+## 2.0.11
+
+Wed, 24 Jul 2024 16:11:28 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.2
+
+## 2.0.10
+
+Tue, 23 Jul 2024 21:56:35 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.1
+
+## 2.0.9
+
+Tue, 23 Jul 2024 20:28:50 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v30.1.0
 
 ## 2.0.8